+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Intel add IP and URL services with lookup endpoint
+
 
 ## [1.0.2] - 2022-12-23